+++ src/database/mod.rs
@@ -4,6 +4,10 @@ pub mod databasetype;
 pub mod datatype;
 pub mod migrations;
 pub mod schema;
+pub mod statement;
+
+#[cfg(test)]
+mod statement_test;
 
 pub use connect::*;
 pub use database::*;
@@ -11,3 +15,4 @@ pub use databasetype::*;
 pub use datatype::*;
 pub use migrations::*;
 pub use schema::*;
+pub use statement::*;